move cookie checks into App, route server paths explicitly
