+++ server/middleware/renderer.js
@@ -9,7 +9,6 @@ const path = require("path");
 const fs = require("fs");
 
 export default (req, res, next) => {
-  console.log("\n\nreq:", req);
   // point to the html file created by CRA's build tool
   const filePath = path.resolve(__dirname, "..", "..", "build", "index.html");
 
@@ -21,7 +20,7 @@ export default (req, res, next) => {
 
     // render the app as a string
     const html = ReactDOMServer.renderToString(
-      <Router context={{}} location={req.url}>
+      <Router context={{}} location={req._parsedUrl}>
         <App />
       </Router>
     );
